removing userTimeline test and keeping a user session test
